move preproc function to step0.rs
